[front] - refactor: move SlackChannelPatternInput to poke directory
 - Renamed file to reflect its specific use within the 'poke' feature module
 - Updated imports in data_sources index page to reflect the new file location
